+++ src/main/java/com/example/gatherplan/appointment/enums/AppointmentState.java
@@ -1,0 +1,8 @@
+package com.example.gatherplan.appointment.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum AppointmentState {
+    CONFIRMED, UNCONFIRMED
+}
