+++ src/main/java/com/example/gatherplan/appointment/enums/CandidateTimeType.java
@@ -1,0 +1,8 @@
+package com.example.gatherplan.appointment.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum CandidateTimeType {
+    CUSTOM,SECTION
+}
