+++ src/main/java/com/example/gatherplan/appointment/repository/AppointmentRepository.java
@@ -1,0 +1,4 @@
+package com.example.gatherplan.appointment.repository;
+
+public interface AppointmentRepository {
+}
