+++ src/main/java/com/example/gatherplan/appointment/repository/entity/EmailAuth.java
@@ -1,0 +1,27 @@
+package com.example.gatherplan.appointment.repository.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+
+@Entity
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class EmailAuth {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String userEmail;
+    private String authCode;
+    private LocalDateTime createdTime;
+    private LocalDateTime expiredTime;
+
+}
