+++ src/main/java/com/example/gatherplan/appointment/repository/entity/Member.java
@@ -1,0 +1,34 @@
+package com.example.gatherplan.appointment.repository.entity;
+
+import com.example.gatherplan.appointment.enums.UserAuthType;
+import com.example.gatherplan.appointment.enums.UserType;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.util.List;
+
+@Entity
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class Member {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "MEMBER_ID")
+    private Long id;
+
+    private String name;
+    private String email;
+    private String password;
+
+    @Enumerated(EnumType.STRING)
+    private UserType userType;
+
+    @Enumerated(EnumType.STRING)
+    private UserAuthType userAuthType;
+
+    @OneToMany(mappedBy = "member")
+    private List<ParticipationEntity> participationEntities;
+}
