+++ src/main/java/com/example/gatherplan/appointment/service/AppointmentService.java
@@ -1,0 +1,4 @@
+package com.example.gatherplan.appointment.service;
+
+public interface AppointmentService {
+}
