+++ src/main/java/com/example/gatherplan/appointment/service/MemberService.java
@@ -1,0 +1,9 @@
+package com.example.gatherplan.appointment.service;
+
+import com.example.gatherplan.appointment.dto.LocalJoinFormDto;
+
+public interface MemberService {
+    void sendAuthCodeProcess(String email);
+
+    void validateLocalJoinFormProcess(LocalJoinFormDto localJoinFormDto);
+}
