+++ src/main/java/com/example/gatherplan/appointment/validation/LocalJoinFormValidationSequence.java
@@ -1,0 +1,10 @@
+package com.example.gatherplan.appointment.validation;
+
+import com.example.gatherplan.common.validation.*;
+import jakarta.validation.GroupSequence;
+
+@GroupSequence({NotBlankEmail.class, PatternCheckEmail.class, NotBlankName.class,
+        SizeCheckName.class , NotBlankPassword.class, SizeCheckPassword.class,
+        PatternCheckPassword.class})
+public interface LocalJoinFormValidationSequence {
+}
