--- src/main/java/com/example/gatherplan/common/dto/response/SuccessResponse.java
@@ -1,17 +1,0 @@
-package com.example.gatherplan.common.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class SuccessResponse {
-
-    private boolean success;
-
-    public static SuccessResponse of() {
-        return SuccessResponse.builder()
-                .success(true)
-                .build();
-    }
-}
