+++ src/main/java/com/example/gatherplan/common/exception/AuthenticationFailException.java
@@ -1,7 +1,11 @@
 package com.example.gatherplan.common.exception;
 
 public class AuthenticationFailException extends BusinessException{
-    public AuthenticationFailException(ExceptionInfo exceptionInfo, String message) {
-        super(exceptionInfo, message);
+    public AuthenticationFailException(ErrorCode errorCode, String message) {
+        super(errorCode, message);
+    }
+
+    public AuthenticationFailException(ErrorCode errorCode) {
+        super(errorCode);
     }
 }
