+++ src/main/java/com/example/gatherplan/common/exception/BusinessException.java
@@ -4,11 +4,16 @@ import lombok.Getter;
 
 @Getter
 public class BusinessException extends RuntimeException{
-    private final ExceptionInfo exceptionInfo;
-    private final String customMessage;
+    private final ErrorCode errorCode;
+    private final String message;
 
-    public BusinessException(ExceptionInfo exceptionInfo, String message) {
-        this.exceptionInfo = exceptionInfo;
-        this.customMessage = message;
+    public BusinessException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+        this.message = errorCode.getMessage();
+    }
+    public BusinessException(ErrorCode errorCode, String customMessage) {
+        this.errorCode = errorCode;
+        this.message = customMessage;
     }
 }
