--- src/main/java/com/example/gatherplan/common/exception/ResourceConflictException.java
@@ -1,7 +1,0 @@
-package com.example.gatherplan.common.exception;
-
-public class ResourceConflictException extends BusinessException{
-    public ResourceConflictException(ExceptionInfo exceptionInfo, String message) {
-        super(exceptionInfo, message);
-    }
-}
