--- src/main/java/com/example/gatherplan/common/exception/ResourceNotFoundException.java
@@ -1,7 +1,0 @@
-package com.example.gatherplan.common.exception;
-
-public class ResourceNotFoundException  extends BusinessException{
-    public ResourceNotFoundException(ExceptionInfo exceptionInfo, String message) {
-        super(exceptionInfo, message);
-    }
-}
