+++ src/main/java/com/example/gatherplan/common/vo/response/ListResponse.java
@@ -1,0 +1,25 @@
+package com.example.gatherplan.common.vo.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ListResponse<T> {
+
+    private MetaData metaData;
+    private T data;
+
+    public ListResponse(MetaData metaData, T data) {
+        this.metaData = metaData;
+        this.data = data;
+    }
+
+    public static <T> ListResponse<T> of(MetaData metaData, T data) {
+        return ListResponse.<T>builder()
+                .metaData(metaData)
+                .data(data)
+                .build();
+    }
+
+}
