--- src/main/java/com/example/gatherplan/domain/embedded/Address.java
@@ -1,17 +1,0 @@
-package com.example.gatherplan.domain.embedded;
-
-import jakarta.persistence.Embeddable;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Embeddable
-@Getter
-@Setter
-@NoArgsConstructor
-public class Address {
-    private String city;
-    private String district;
-    private String street;
-    private String buildingNumber;
-}
