--- src/main/java/com/example/gatherplan/domain/enums/AppointmentState.java
@@ -1,8 +1,0 @@
-package com.example.gatherplan.domain.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum AppointmentState {
-    CONFIRMED, UNCONFIRMED
-}
