--- src/main/java/com/example/gatherplan/domain/enums/CandidateTimeType.java
@@ -1,8 +1,0 @@
-package com.example.gatherplan.domain.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum CandidateTimeType {
-    CUSTOM,SECTION
-}
