--- src/main/java/com/example/gatherplan/feature/join/dto/LocalJoinEmailFormDto.java
@@ -1,20 +1,0 @@
-package com.example.gatherplan.feature.join.dto;
-
-import com.example.gatherplan.common.validation.NotBlankEmail;
-import com.example.gatherplan.common.validation.PatternCheckEmail;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class LocalJoinEmailFormDto {
-    @NotBlank(message = "이메일은 공백이 될 수 없습니다.", groups = NotBlankEmail.class)
-    @Email(message = "이메일 형식이 맞지 않습니다.", groups = PatternCheckEmail.class)
-    private String email;
-}
