--- src/main/java/com/example/gatherplan/feature/join/service/LocalJoinService.java
@@ -1,10 +1,0 @@
-package com.example.gatherplan.feature.join.service;
-
-import com.example.gatherplan.feature.join.dto.LocalJoinEmailFormDto;
-import com.example.gatherplan.feature.join.dto.LocalJoinFormDto;
-
-public interface LocalJoinService {
-    void sendAuthCodeProcess(LocalJoinEmailFormDto localJoinEmailFormDto);
-
-    void validateLocalJoinFormProcess(LocalJoinFormDto localJoinFormDto);
-}
